问题修复 - 修复实用拓展菜田的收获错误问题 - 修复读取模型文件的崩溃问题 - 修复光影下渲染稻草人头颅的变色问题 - 移除符卡栏 - 修复种菜、种甘蔗可能会出现的种在水里的问题 - 服务端无模型包时，不发送更新信息

+++ src/main/java/com/github/tartaricacid/touhoulittlemaid/client/model/pojo/GeometryModel.java
@@ -15,10 +15,10 @@ public class GeometryModel {
     private int texturewidth;
 
     @SerializedName("visible_bounds_height")
-    private int visibleBoundsHeight;
+    private float visibleBoundsHeight;
 
     @SerializedName("visible_bounds_width")
-    private int visibleBoundsWidth;
+    private float visibleBoundsWidth;
 
     @SerializedName("visible_bounds_offset")
     private List<Float> visibleBoundsOffset;
@@ -35,11 +35,11 @@ public class GeometryModel {
         return texturewidth;
     }
 
-    public int getVisibleBoundsHeight() {
+    public float getVisibleBoundsHeight() {
         return visibleBoundsHeight;
     }
 
-    public int getVisibleBoundsWidth() {
+    public float getVisibleBoundsWidth() {
         return visibleBoundsWidth;
     }
